Fixed an issue with realignment of non-aligned buffers
When re-assembling packets for transfer in a chained buffer, sometimes the buffer address for the first packet wouldn't be aligned as the ICMP fragmentation code passes in the buffer pointer to the payload. When this happens, the unaligned bytes in the packet will get chopped off and the transfer out will fail. This usually appeared as packet corruption in the ICMP reply when packets are fragmented. This fix isn't perfect, but works for the LPC devices that use 32-bit alignment in copied and zero-copy modes.