Several small enhancements, GNU support
Changed an error message for the copied mode EMAC driver to output EMAC status on an error
Update the way the UART debug redirect code is included in the build based on the definition on LWIP_DEBUG
Added GNU support for http_raw and echo projects. dram_zc version of echo works, but other projects might need some work still.